feat: revert MB sections to their prev update behavior
